move components up one level and update paths
