+++ client/src/App.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { ThemeProvider } from 'styled-components';
 import theme from './styled-components/Theme';
-import PreLogin from './components/grow-sprint-2/PreLogin';
-import MainContainer from './components/grow-sprint-2/MainContainer';
-import Header from './components/grow-sprint-2/Header';
+import PreLogin from './components/PreLogin';
+import MainContainer from './components/MainContainer';
+import Header from './components/Header';
 
 import './scss/app.scss';
