+++ client/src/components/Result.tsx
